Use Garden Type
Instead of querying for fields in `site` that might not exist (ex: if
an option isn't set by a user and uses the default), this PR creates a
`Garden` type used the schema customization APIs. Then we set a
specific, known, ID on the node we insert so we can fetch it later and
not have our queries fail due to missing data.